+++ src/Maybe/and.ts
@@ -1,5 +1,5 @@
 import { isNotNullAndUndefined, Maybe } from './Maybe';
 
-export function andMaybe<T, U>(a: Maybe<T>, b: Maybe<U>): Maybe<U> {
+export function andForMaybe<T, U>(a: Maybe<T>, b: Maybe<U>): Maybe<U> {
     return isNotNullAndUndefined(a) ? b : a;
 }