+++ src/Nullable/or.ts
@@ -1,5 +1,5 @@
 import { Nullable } from './Nullable';
 
-export function orNullable<T>(a: Nullable<T>, b: Nullable<T>): Nullable<T> {
+export function orForNullable<T>(a: Nullable<T>, b: Nullable<T>): Nullable<T> {
     return (a !== null) ? a : b;
 }