+++ src/PlainOption/index.ts
@@ -17,7 +17,7 @@ export { expectIsSome as expect } from './expect';
 export { mapForOption as map } from './map';
 export { mapOrForOption as mapOr } from './mapOr';
 export { mapOrElseForOption as mapOrElse } from './mapOrElse';
-export { orOption as or } from './or';
+export { orForOption as or } from './or';
 export { orElseForOption as orElse } from './orElse';
 export { unwrapOption as unwrap } from './unwrap';
 export { unwrapOrFromOption as unwrapOr } from './unwrapOr';