+++ src/Undefinable/and.ts
@@ -1,5 +1,5 @@
 import { Undefinable, isNotUndefined } from './Undefinable';
 
-export function andUndefinable<T, U>(a: Undefinable<T>, b: Undefinable<U>): Undefinable<U> {
+export function andForUndefinable<T, U>(a: Undefinable<T>, b: Undefinable<U>): Undefinable<U> {
     return isNotUndefined(a) ? b : a;
 }