+++ src/Undefinable/or.ts
@@ -1,5 +1,5 @@
 import { Undefinable, isNotUndefined } from './Undefinable';
 
-export function orUndefinable<T>(a: Undefinable<T>, b: Undefinable<T>): Undefinable<T> {
+export function orForUndefinable<T>(a: Undefinable<T>, b: Undefinable<T>): Undefinable<T> {
     return isNotUndefined(a) ? a : b;
 }